+++ tedi/components/filter/index.ts
@@ -1,0 +1,9 @@
+export {
+  FilterComponent,
+  FilterVariant,
+  FilterSize,
+  FilterOption,
+} from "./filter.component";
+export { FilterContentDirective } from "./filter-content.directive";
+export { FilterPrependDirective } from "./filter-prepend.directive";
+export { FilterGroupComponent } from "./filter-group.component";