+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# [6.4.0-rc.5](https://github.com/TEDI-Design-System/angular/compare/angular-6.4.0-rc.4...angular-6.4.0-rc.5) (2026-05-22)
+
+
+### Features
+
+* **filter,status-indicator:** new filter component, status-indicator [#329](https://github.com/TEDI-Design-System/angular/issues/329) ([#396](https://github.com/TEDI-Design-System/angular/issues/396)) ([99e93b3](https://github.com/TEDI-Design-System/angular/commit/99e93b37830a89ea24fc45b706e09ec677433e80)), closes [#222](https://github.com/TEDI-Design-System/angular/issues/222) [#222](https://github.com/TEDI-Design-System/angular/issues/222) [#222](https://github.com/TEDI-Design-System/angular/issues/222) [#7](https://github.com/TEDI-Design-System/angular/issues/7) [#7](https://github.com/TEDI-Design-System/angular/issues/7) [#222](https://github.com/TEDI-Design-System/angular/issues/222) [#222](https://github.com/TEDI-Design-System/angular/issues/222)
+
+# [6.4.0-rc.4](https://github.com/TEDI-Design-System/angular/compare/angular-6.4.0-rc.3...angular-6.4.0-rc.4) (2026-05-20)
+
+
+### Bug Fixes
+
+* **alert:** renamed error to danger in storybook type options array [#438](https://github.com/TEDI-Design-System/angular/issues/438) ([#444](https://github.com/TEDI-Design-System/angular/issues/444)) ([23acf9c](https://github.com/TEDI-Design-System/angular/commit/23acf9c3d3a4bb67f80ed06beccbb31010cc7e61))
+
+# [6.4.0-rc.3](https://github.com/TEDI-Design-System/angular/compare/angular-6.4.0-rc.2...angular-6.4.0-rc.3) (2026-05-18)
+
+
+### Bug Fixes
+
+* **collapse:** add button type to prevent form submission [#434](https://github.com/TEDI-Design-System/angular/issues/434) ([#441](https://github.com/TEDI-Design-System/angular/issues/441)) ([2e13aef](https://github.com/TEDI-Design-System/angular/commit/2e13aefdd60145b58b2e1053b09fa89a9a1a391c))
+
+
+### Reverts
+
+* **horizontal-stepper:** undo premature rc commits [#405](https://github.com/TEDI-Design-System/angular/issues/405) ([#436](https://github.com/TEDI-Design-System/angular/issues/436)) ([a575670](https://github.com/TEDI-Design-System/angular/commit/a57567051e7a522ab1e691b46c5ef4d52f599f0f))
+
 # [6.4.0-rc.2](https://github.com/TEDI-Design-System/angular/compare/angular-6.4.0-rc.1...angular-6.4.0-rc.2) (2026-05-08)